[WRFE-55](refactor): 홈 폴더구조 리팩토링
