+++ apps/front/wraffle-webview/app/page.tsx
@@ -1,10 +1,10 @@
-import {RecentRaffles} from '@/entities/banner/constants';
-import {categories} from '@/entities/category/constants';
-import {recommendItems} from '@/entities/raffle/recommend/constants';
-import Banner from '@/features/manage-raffle/ui/Banner';
+import {categories} from '@/entities/category';
+import {RecentRaffles} from '@/features/manage-raffle/config';
+import {Banner} from '@/features/manage-raffle/ui';
 import {Header} from '@/shared/ui';
-import CategoryButtons from '@/widgets/category-list/ui/CategoryButtons';
-import RecommendList from '@/widgets/recommend-list/ui/RecommendList';
+import {CategoryButtons} from '@/widgets/category-list/ui';
+import {recommendItems} from '@/widgets/recommend-list/config';
+import {RecommendList} from '@/widgets/recommend-list/ui';
 import {BottomNavigation, Icon} from '@wraffle/ui';
 
 export default function Home() {
