+++ apps/front/wraffle-webview/src/entities/raffle/const.ts
@@ -1,3 +1,1 @@
 export const RAFFLE_TYPE = ['RAFFLE', 'EVENT'] as const;
-
-export type RaffleType = (typeof RAFFLE_TYPE)[number];
