+++ apps/front/wraffle-webview/src/features/manage-raffle/model/recommend.ts
@@ -1,0 +1,6 @@
+import type {Raffle, RaffleType} from '@/entities/raffle';
+
+export interface RecommendRaffle extends Raffle {
+  type: RaffleType;
+  isBookmarked: boolean;
+}
