+++ apps/front/wraffle-webview/src/widgets/category-list/config/const.ts
@@ -1,0 +1,76 @@
+import type {CategoryItem} from '@/entities/category';
+
+export const categories: CategoryItem[] = [
+  {
+    id: 1,
+    name: '패션',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 2,
+    name: '가전 디지털',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 3,
+    name: '전자제품',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 4,
+    name: '식품',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 5,
+    name: '의류',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 6,
+    name: '생활',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 11,
+    name: '패션',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 12,
+    name: '가전 디지털',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 13,
+    name: '전자 제품',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 14,
+    name: '식품',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 15,
+    name: '의류',
+    parentId: null,
+    depth: 1,
+  },
+  {
+    id: 16,
+    name: '생활',
+    parentId: null,
+    depth: 1,
+  },
+];
