+++ apps/front/wraffle-webview/src/widgets/category-list/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './CategoryButtons';
