+++ apps/front/wraffle-webview/src/widgets/recommend-list/config/index.ts
@@ -1,0 +1,1 @@
+export * from './const';
