+++ apps/front/wraffle-webview/src/widgets/recommend-list/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './RecommendList';
